chore(legacy-sweep): tier 3 — rename internal types, fix mock, add v1.0.0 cutoff (#622)
§4.18 Phase A Tier 3 INVESTIGATE items resolved:

- packages/core/src/types/legacy.ts -> internal.ts: file is misnamed.
  Audit confirmed all 11 exports are @internal namespace markers (Block,
  CheckboxField, BlocksField, RevealUIBlock, ClientConfig,
  ClientCollectionConfig, RevealUITraverseFieldsArgs, etc.) — no
  deprecated markers, no back-compat language. The file's own JSDoc
  already declared @module @revealui/core/types/internal. Rename matches
  reality. 2 consumers updated (fieldTraversal.ts, types/index.ts).

- ticket-comments mock: add missing authorId + updatedAt fields per
  schema in packages/db/src/schema/tickets.ts:239-264. (code-provenance
  mock verified clean — all 19 fields already present, no drift.)

- revealui shell v1.0.0 cutoff: add discoverable TODO comment in
  packages/cli/src/cli.ts above the deprecated alias + update the
  command description to mention the v1.0.0 removal date. Per founder
  direction 2026-04-27.

Phase A Tier 3 now closed. Tier 4 (re-scan @revealui/ai/mcp/core
barrels for orphaned re-exports) remains outstanding.
